Ultima modificacion

+++ DesafioEntregable/ProductoHandler.cs
@@ -28,8 +28,8 @@ namespace DesafioEntregable
                                 Producto producto = new Producto();
                                 producto.Id = Convert.ToInt32(dataReader["Id"]);
                                 producto.Descripcion = dataReader["Descripciones"].ToString();
-                                producto.Costo = Convert.ToInt32(dataReader["Costo"]);
-                                producto.PrecioVenta = Convert.ToInt32(dataReader["PrecioVenta"]);
+                                producto.Costo = Convert.ToDouble(dataReader["Costo"]);
+                                producto.PrecioVenta = Convert.ToDouble(dataReader["PrecioVenta"]);
                                 producto.Stock = Convert.ToInt32(dataReader["Stock"]);
                                 producto.IdUsuario = Convert.ToInt32(dataReader["IdUsuario"]);
